Fixes for MSVC 2008 and 2010
MSVC pre-2015 is not C99/C11/C++11/C++14 compliant and therefore
does not contain a lot of the standard library math functions
that are present in compilers such as clang and gcc. Luckily,
boost provides implementations of all of these functions, so
we can just important them if we are using MSVC. However,
the one issue is FMA - which is just templated back to x*y+z,
which obviously negates the advantage of the function in the
first place.