Move windows tests to later in the build:
They'll still always run on a PR but this way they won't run until after the unit tests have already passed in some other configuration.
As is, unit tests could fail and we'd still queue up and take up an hour of time on Windows even though we tend not to learn something new with Windows tests.
Windows is a big bottleneck right now.