Make fixed point algebra solver variadic.
This commit updates the fixed point solver so that variadic arguments
can be passed to it. It also uses a more efficient method of propagating
adjoints, in line with the other solvers. This required restructuring
some aspects of the code slightly; it is now organized to mimic the code
for the Newton solver as closely as possible.

This has resulted in a number of tests no longer being applicable. I've
commented them out for now, and will go through more carefully in the
future.